+++ example/CMakeLists.txt
@@ -2,4 +2,3 @@ add_subdirectory(miniapp)
 add_subdirectory(generators)
 add_subdirectory(brunel)
 add_subdirectory(bench)
-add_subdirectory(threadbench)
